requirements.txt
flask>=2.3
pystray>=0.19
pillow>=10.0
scikit-learn>=1.3
